+++ sei-db/state_db/sc/migration/passthrough_router.go
@@ -71,4 +71,4 @@ func (p *PassthroughRouter) GetProof(store string, key []byte) (*ics23.Commitmen
 
 // SetMigrationBatchSize is a no-op: a passthrough router targets a single
 // backend and performs no data migration.
-func (p *PassthroughRouter) SetMigrationBatchSize(int) {}
+func (p *PassthroughRouter) SetMigrationBatchSize(uint64) {}
